+++ src/YFastTrieInterface.cpp
@@ -1,11 +1,11 @@
 #include "YFastTrieInterface.hpp"
 
-YFastTrieInterface::YFastTrieInterface(std::vector<long long int>& input, long long int universe)
+YFastTrieInterface::YFastTrieInterface(std::vector<int>& input, int universe)
 {
   tree_ = new YFastTrie(universe, input);
 }
 
-YFastTrieInterface::YFastTrieInterface(std::vector<long long int>& input, long long int universe, int state)
+YFastTrieInterface::YFastTrieInterface(std::vector<int>& input, int universe, int state)
 {
   tree_ = new YFastTrie(universe, input, state);
 }
@@ -20,17 +20,17 @@ YFastTrieInterface::~YFastTrieInterface()
   }
 }
 
-long long int YFastTrieInterface::predecessor(long long int x)
+int YFastTrieInterface::predecessor(int x)
 {
   return tree_ -> predecessor(x);
 }
 
-long long int YFastTrieInterface::successor(long long int x)
+int YFastTrieInterface::successor(int x)
 {
   return tree_ -> successor(x);
 }
 
-bool YFastTrieInterface::member(long long int x)
+bool YFastTrieInterface::member(int x)
 {
   return tree_ -> member(x);
 }
